aicp: changelog: Only rebuild changelog when needed, and cleanup
If we have built the changelog on a prior build and it is still up to
date, skip building the changelog again.

In order to detect whether the changelog is up to date, we write a
revisioned manifest.  If the manifest exists (from a prior run), it is
not newer than the changelog, and it matches the current repo state,
then the changelog is up to date and we don't need to do anything.

When we do need to write a changelog, we remove the old manifest, write
the changelog, then update the manifest.  This ensures that if any repo
command fails, the manifest will not exist and the changelog will be
written again on the next run.

Also do some cleanup:
 * Use 'set -e' to detect failure in repo commands.
 * Write changelog to $OUT, not the build topdir.
 * Exit early when USER_BUILD_NO_CHANGELOG is set.
 * Use quotes around variable references.
 * Make all local shell variables lower case.
 * Do not export variables needlessly.

Finally note that most or all of this could have been done in directly
in the core build system as make rules.  But let's leave that change
for another time.

Change-Id: Ib390007cda3f8148e96ceeca23db97bde3fcc15c